+++ include/GlobalManager.hpp
@@ -14,7 +14,7 @@ public:
     int GetNchannel() const;
     int GetNcol() const;
     double GetTimeRefAcc() const;
-    double GetPeakTolerance() const;
+    float GetPeakTolerance() const;
     void ApplyConfig(int run); 
     // You can add additional getters as needed.
 private:
