+++ include/ReferenceManager.hpp
@@ -19,8 +19,8 @@ public:
     const ROOT::Math::Interpolator* GetInterpolator(int block) const;
     TF1* GetFitter(int block) const;
     
-    std::map<int,double> GetTimeRefs() const;
-    std::map<int,double> GettdcOffsets() const;
+    std::map<int,float> GetTimeRefs() const;
+    std::map<int,float> GettdcOffsets() const;
     double GetTimeMean() const;
     double GetTimeMean2() const;
     void ApplyConfig(int _run);
@@ -32,8 +32,8 @@ private:
     // Maps to store interpolators and fitters.
     std::map<int, std::unique_ptr<ROOT::Math::Interpolator>> interpolators;
     // TODO: Get 
-    std::map<int,double> timeRefs;
-    std::map<int,double> tdcOffsets;
+    std::map<int,float> timeRefs;
+    std::map<int,float> tdcOffsets;
     std::map<int, std::unique_ptr<TF1>> fitters;
 };
